mostly bug fixes

+++ src/multiseason.h
@@ -12,12 +12,12 @@ struct {
 void INIT_MULTISEASON(float *parList) ;
 void init_multiseason__(float *parList);
 
-void GET_MULTISEASON(char *CCID, int NOBS, float *MJD, 
-		     float *FLUX, float *FLUXERR,
-		     int *REJECT, int *NSEASON, float *CHI2RED, 
-		     float *MJDMIN, float *MJDMAX, float *AVGFLUX);
+void GET_MULTISEASON(char *CCID, int NOBS, double *MJD, 
+		     double *FLUX, double *FLUXERR,
+		     int *REJECT, int *NSEASON, double *CHI2RED, 
+		     double *MJDMIN, double *MJDMAX, double *AVGFLUX);
 
-void get_multiseason__(char *CCID, int *NOBS, float *MJD, 
-		       float *FLUX, float *FLUXERR,
-		       int *REJECT, int *NSEASON, float *CHI2RED,
-		       float *MJDMIN, float *MJDMAX, float *AVGFLUX ) ;
+void get_multiseason__(char *CCID, int *NOBS, double *MJD, 
+		       double *FLUX, double *FLUXERR,
+		       int *REJECT, int *NSEASON, double *CHI2RED,
+		       double *MJDMIN, double *MJDMAX, double *AVGFLUX ) ;
